+++ design-system/react/src/components/Button/defs.ts
@@ -1,15 +1,16 @@
 import type { LayerIntent, LayerVariant } from '@fuel-ui/css';
+import type { PressEvents } from '@react-types/shared';
 import type { AriaButtonProps } from 'react-aria';
 import type { CreateComponent } from '~/utils';
 import type { Components } from '~/utils/components-list';
 
-import type { HTMLProps, PressProps } from '../../utils/types';
+import type { HTMLProps } from '../../utils/types';
 import type { IconProps } from '../Icon';
 
 export type ButtonVariants = 'solid' | 'outlined' | 'ghost' | 'link';
 export type ButtonSizes = 'xs' | 'sm' | 'md' | 'lg';
 
-export type ButtonBaseProps = {
+export type ButtonCustomProps = {
   size?: ButtonSizes;
   intent?: LayerIntent;
   variant?: LayerVariant;
@@ -23,10 +24,11 @@ export type ButtonBaseProps = {
   isDisabled?: boolean;
 };
 
-export type ButtonProps = Omit<HTMLProps['button'], 'onClick'> &
-  AriaButtonProps<'button'> &
-  PressProps &
-  ButtonBaseProps & {
+export type ButtonBaseProps = HTMLProps['button'] &
+  Omit<AriaButtonProps<'button'>, keyof PressEvents | 'onClick'>;
+
+export type ButtonProps = ButtonBaseProps &
+  ButtonCustomProps & {
     justIcon?: boolean;
     isLink?: boolean;
   };