+++ design-system/react/src/components/ButtonLink/ButtonLink.stories.tsx
@@ -16,7 +16,7 @@ const meta: Meta<typeof ButtonLink> = {
     children: {
       control: false,
     },
-    onPress: {
+    onClick: {
       control: false,
     },
     size: {