+++ design-system/react/src/components/Card/Card.tsx
@@ -1,6 +1,6 @@
 import { mergeProps } from '@react-aria/utils';
 import { useStyles } from '~/hooks';
-import { useOnPress } from '~/hooks/useOnPress';
+import { useOnClick } from '~/hooks/useOnClick';
 import { Components } from '~/utils/components-list';
 
 import {
@@ -17,16 +17,18 @@ import { styles } from './styles';
 
 const _Card = _unstable_createComponent<CardDef>(
   Components.Card,
-  ({ as = 'article', ref, variant, withDividers, children, ...props }) => {
+  ({
+    as = 'article',
+    ref,
+    variant,
+    withDividers,
+    onClick,
+    children,
+    ...props
+  }) => {
     const classes = useStyles(styles, props);
-    const { buttonProps } = useOnPress<CardDef['props'], CardDef['element']>(
-      props,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      ref as any,
-      { elementType: as },
-    );
-
-    const isClickable = Boolean(props.onPress);
+    const { buttonProps } = useOnClick(ref, { onClick, elementType: as });
+    const isClickable = Boolean(onClick);
     const elementProps = {
       ...props,
       ref,