+++ design-system/react/CHANGELOG.md
@@ -429,7 +429,7 @@
   ```jsx
   import { Button, toast } from '@fuel-ui/react';
   function App() {
-    return <Button onPress={() => toast('Hello world!')}>Show toast</Button>;
+    return <Button onClick={() => toast('Hello world!')}>Show toast</Button>;
   }
   ```